+++ mars/stn/stn.h
@@ -414,7 +414,7 @@ struct IPPortItem {
 extern bool MakesureAuthed(const std::string& _host, const std::string& _user_id);
 
 //流量统计
-extern void TrafficData(ssize_t _send, ssize_t _recv);
+extern void TrafficData(int64_t _send, int64_t _recv);
 
 //底层询问上层该host对应的ip列表
 extern std::vector<std::string> OnNewDns(const std::string& _host, bool _longlink_host);
