+++ mars/stn/stn_callback_bridge.h
@@ -31,7 +31,7 @@ class StnCallbackBridge {
 
     virtual bool MakesureAuthed(const std::string& _host, const std::string& _user_id);
 
-    virtual void TrafficData(ssize_t _send, ssize_t _recv);
+    virtual void TrafficData(int64_t _send, int64_t _recv);
 
     virtual std::vector<std::string> OnNewDns(const std::string& _host,
                                               bool _longlink_host,
