+++ mars/stn/stn_logic.h
@@ -50,7 +50,7 @@ public:
     virtual bool MakesureAuthed(const std::string& _host, const std::string& _user_id) = 0;
 
     //流量统计
-    virtual void TrafficData(ssize_t _send, ssize_t _recv) = 0;
+    virtual void TrafficData(int64_t _send, int64_t _recv) = 0;
 
     //底层询问上层该host对应的ip列表
     virtual std::vector<std::string> OnNewDns(const std::string& host, bool _longlink_host) = 0;
@@ -184,7 +184,7 @@ StnCallbackBridge* GetStnCallbackBridge();
 extern bool MakesureAuthed(const std::string& _host, const std::string& _user_id);
 
 //流量统计
-extern void TrafficData(ssize_t _send, ssize_t _recv);
+extern void TrafficData(int64_t _send, int64_t _recv);
 
 //底层询问上层该host对应的ip列表
 extern std::vector<std::string> OnNewDns(const std::string& _host,
