+++ mars/stn/stn_manager.h
@@ -40,7 +40,7 @@ class StnManager : public mars::boot::BaseManager {
     void OnSingalCrash(int _sig);
     void OnExceptionCrash();
     void OnNetworkChange(void (*pre_change)());
-    void OnNetworkDataChange(const char* _tag, ssize_t _send, ssize_t _recv);
+    void OnNetworkDataChange(const char* _tag, int64_t _send, int64_t _recv);
 #ifdef ANDROID
     void OnAlarm(int64_t _id);
 #endif
@@ -58,7 +58,7 @@ class StnManager : public mars::boot::BaseManager {
     bool MakesureAuthed(const std::string& _host, const std::string& _user_id);
 
     // 流量统计
-    void TrafficData(ssize_t _send, ssize_t _recv);
+    void TrafficData(int64_t _send, int64_t _recv);
 
     // 底层询问上层该host对应的ip列表
     std::vector<std::string> OnNewDns(const std::string& _host,
